Bugfix: ValueError: If `preds` and `target` are of shape (N, ...) and `preds` are floats, `target` should be binary.

+++ src/models/segformer/make_lightning.py
@@ -126,7 +126,7 @@ class SegSubLightning(pl.LightningModule):
 
     def validation_log(self, batch_idx, outputs, inputs):
         outputs = self.logits_to_labels(outputs)
-        self.metrics.update(outputs.float(), inputs['labels'])
+        self.metrics.update(outputs, inputs['labels'])
 
         if batch_idx == 0:
             self.log_image(inputs, outputs)
